Feat: Recipe details and showed and hid by clicking the recipe name.
Now when the page loads all recipes' details will be hidden.
Visibility of a crecipe's details can be toggled by clicking the its name.
The visibility is controlled by states (jQuery and css not used to do this).
getInitialState is added to RecipeList class. handleClick in RecipeList is modified as well.