Feat: New Recipe form now functional.
The user can now add a new recipe by clicking New Recipe button and filling in information.
newRecipeSubmit method was added to Box to handle display of the new recipe.
cancelNewRecipe method was also added to Box to handle clicking the cancel button (hiding the form).
In NewRecipeForm, two methods were added: submit to handle submission of the form (newRecipeSubmit is called here).
The second method is handleCancel to handle clicking cancel (cancelNewRecipe is called from this method).
Next step is to enable editing and deletion of an existing recipe.