Feat: Editing recipes now available.
The user can now edit exisitng recipes. The following was changed in the code:
editSubmit method added to Box.
editSubmit is added to RecipeList.
getInitialState, handleEdit, onCancel and onSubmit were added to RecipeDetails.
EditRecipeForm component added with methods handleCancel, submit and render.
submit handlers are added to four components because the new recipe information are passed through from EditRecipeform to Box.
Next is activate delete button to delete a recipe.