Bug: Fixed bug with deleting some recipes.
Up to this point, to delete a recipe, the process was done by splicing recipes at the key of the recipe.
This generated a bug because when an intermediate recipe is deleted, the last element's key will be larger than the last index in recipes.
So deleteHandler in Box was modified to detec the index of the recipe and use the index for the splice.
A new state c is added to Box as well. This state holds the value of the key of next new recipe.
The key was assigned based on the length of recipes array. This is Ok if we don't delete recipes.
If we try to add a recipe after deleting an intermediate one, we'll end up with recipes having same key.
So now using c, we guarantee having a unique key for each recipe.