Feat: Data are now saved in browser storage.
Whenever the user edits, saves or deletes a recipe. All recipes and keys are saved in local storage.
If statement added to getInitialState to check and load data if local storage has any.
saveData method added to Box to save keys and recipes.
A call to saveData was placed in deleteHandler, editSubmit and newRecipeSubmit metohds.