Bug: Fixed bug which prevented editing some recipes.
After deleting a recipe, some recipes become 'uneditable'. The problem was with keys of the recipes.
After deleting a recipe, the index of some recipes change in the recipes array.
key doesn't refer to the actual index of the recipe in the array.
So now recipes is looped through and the index of the recipe with the specified key is found.
This index is used to edit the recipe.
Application fully funcitonal now and can be submitted. A few tweaks might be added.