Feat: When a recipe is clicked, all other recipes are collapsed.
When a recipe is expanded, its details remain shown until that recipe is clicked again to collapse its details.
Now, If a recipe is clicked, all other recipes are collapsed automatically.
Still, if a recipe is clicked when it's expanded, it will collapse.
Changes were made in getInitialState where keys are now used as property name in intialState.
Changes are also made in handleClick where all states are set to false and state of the clicked recipe is negated.
These methods are in RecipeList.